# cmd/api/docs/swagger.yaml
definitions:
  github_com_fazilnbr_project-workey_pkg_domain.Signup:
    properties:
      countrycode:
        type: string
      otp:
        type: string
      phonenumber:
        type: string
    type: object
  github_com_fazilnbr_project-workey_pkg_domain.UserData:
    properties:
      dob:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      profilephoto:
        type: string
      userId:
        type: integer
    required:
    - profilephoto
    type: object
  github_com_fazilnbr_project-workey_pkg_utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /user/add-profile:
    post:
      operationId: AddProfileAndUpdateMail
      parameters:
      - description: User Data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_domain.UserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: Add User Profile And Update Mail
      tags:
      - User Profile Management
  /user/login-gl:
    get:
      operationId: Authenticate With Google
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: Authenticate With Google
      tags:
      - User Authentication
  /user/sent-otp:
    post:
      operationId: sendOtp
      parameters:
      - description: Mobile Number
        in: body
        name: mobileNumber
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_domain.Signup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
      summary: Send OTP for Users
      tags:
      - User Authentication
  /user/signup-and-login:
    post:
      operationId: SignUp authentication
      parameters:
      - description: Mobile Number And OTP
        in: body
        name: mobileNumberAndOTP
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_domain.Signup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_project-workey_pkg_utils.Response'
      summary: SignUp for users
      tags:
      - User Authentication
swagger: "2.0"